requirements.txt
fastapi
uvicorn[standard]
jinja2
pymysql
python-multipart
scikit-learn
yoyo-migrations